+++ api/store.js
@@ -15,7 +15,7 @@
  *
  * @exports Store
  *
- * @version 2.12.0
+ * @version 2.13.0
  * @author  Shinei Nouzen
  * @license MIT
  * ======= • ======= • ======= • ======= • =======• =======